refactor(MovieApp): Refactor MovieViewModel to use Coroutines
